fix(flake.lock): 🐛 Update catppuccin-userstyles and nixpkgs versions
Updated the locked versions of catppuccin-userstyles and nixpkgs in flake.lock to ensure compatibility with the latest changes. This includes modifications to the lastModified timestamps and narHashes for both dependencies, as well as revisions to their respective repositories.

This change is necessary to maintain the integrity and functionality of the user styles.